calculator_app/requirements.txt
Cerberus==1.3.4
Flask==2.2.2
python_dateutil==2.8.2
